+++ coverage_baseline.txt
@@ -1,0 +1,96 @@
+Name                                        Stmts   Miss  Cover
+---------------------------------------------------------------
+examples/cosim/room/processor.py               74     11    85%
+examples/cosim/thermostat/processor.py         84     14    83%
+examples/prime/factor_search/processor.py      45      1    98%
+examples/prime/factorisation/processor.py      70      1    99%
+examples/simple/abc/processor.py               62      4    94%
+examples/simple/ping/processor.py              96     12    88%
+examples/simple/ping/server.py                108     28    74%
+plugins/dor_default/__init__.py                 2      0   100%
+plugins/dor_default/service.py                337     30    91%
+plugins/rti_aws/__init__.py                     2      0   100%
+plugins/rti_aws/service.py                    327    106    68%
+plugins/rti_docker/__init__.py                  2      0   100%
+plugins/rti_docker/service.py                 248     68    73%
+simaas/__init__.py                              0      0   100%
+simaas/cli/__init__.py                          0      0   100%
+simaas/cli/cmd_dor.py                         311     64    79%
+simaas/cli/cmd_identity.py                    265     88    67%
+simaas/cli/cmd_image.py                       343     58    83%
+simaas/cli/cmd_job_runner.py                  493     56    89%
+simaas/cli/cmd_namespace.py                    95     18    81%
+simaas/cli/cmd_network.py                      21      0   100%
+simaas/cli/cmd_rti.py                         487    163    67%
+simaas/cli/cmd_service.py                     141     67    52%
+simaas/cli/exceptions.py                        5      0   100%
+simaas/cli/helpers.py                         273    139    49%
+simaas/core/__init__.py                         0      0   100%
+simaas/core/assets.py                         118      6    95%
+simaas/core/eckeypair.py                       65      2    97%
+simaas/core/exceptions.py                      32      2    94%
+simaas/core/helpers.py                        114     48    58%
+simaas/core/identity.py                        26      2    92%
+simaas/core/keypair.py                         43      9    79%
+simaas/core/keystore.py                       164     20    88%
+simaas/core/logging.py                         59      3    95%
+simaas/core/processor.py                       56     12    79%
+simaas/core/rsakeypair.py                     119      4    97%
+simaas/core/schemas.py                         19      0   100%
+simaas/decorators.py                           27      0   100%
+simaas/dor/__init__.py                          0      0   100%
+simaas/dor/api.py                              94      1    99%
+simaas/dor/exceptions.py                       14      1    93%
+simaas/dor/protocol.py                        125      4    97%
+simaas/dor/schemas.py                          86      0   100%
+simaas/helpers.py                             199     32    84%
+simaas/meta.py                                  4      0   100%
+simaas/namespace/__init__.py                    0      0   100%
+simaas/namespace/api.py                        15      0   100%
+simaas/namespace/default.py                   115     24    79%
+simaas/namespace/protocol.py                  150     19    87%
+simaas/node/__init__.py                         0      0   100%
+simaas/node/base.py                           194     19    90%
+simaas/node/default.py                         35      0   100%
+simaas/nodedb/__init__.py                       0      0   100%
+simaas/nodedb/api.py                           61      1    98%
+simaas/nodedb/default.py                      234     22    91%
+simaas/nodedb/exceptions.py                    20      5    75%
+simaas/nodedb/protocol.py                     280     44    84%
+simaas/nodedb/schemas.py                       20      0   100%
+simaas/p2p/__init__.py                          0      0   100%
+simaas/p2p/base.py                            130     21    84%
+simaas/p2p/exceptions.py                        8      0   100%
+simaas/p2p/protocol.py                         98     22    78%
+simaas/p2p/service.py                         104      7    93%
+simaas/plugins/__init__.py                      2      0   100%
+simaas/plugins/discovery.py                    64     14    78%
+simaas/rest/__init__.py                         0      0   100%
+simaas/rest/auth.py                           111      3    97%
+simaas/rest/exceptions.py                      20      3    85%
+simaas/rest/proxy.py                          181     45    75%
+simaas/rest/schemas.py                         14      0   100%
+simaas/rest/service.py                         71      4    94%
+simaas/rti/__init__.py                          0      0   100%
+simaas/rti/api.py                             105      3    97%
+simaas/rti/base.py                            379     68    82%
+simaas/rti/exceptions.py                       35      6    83%
+simaas/rti/protocol.py                        137     21    85%
+simaas/rti/schemas.py                         106      0   100%
+simaas/tests/__init__.py                        0      0   100%
+simaas/tests/conftest.py                      382     40    90%
+simaas/tests/test_cli.py                     1497    240    84%
+simaas/tests/test_core.py                     246      0   100%
+simaas/tests/test_example_abc.py              106      7    93%
+simaas/tests/test_example_cosim.py            128     19    85%
+simaas/tests/test_example_ping.py             119      4    97%
+simaas/tests/test_example_primes.py           181     18    90%
+simaas/tests/test_service_dor.py              324      3    99%
+simaas/tests/test_service_namespace.py        152      6    96%
+simaas/tests/test_service_nodedb.py           264      2    99%
+simaas/tests/test_service_p2p.py              201     30    85%
+simaas/tests/test_service_rest.py             157      4    97%
+simaas/tests/test_service_rti.py              548     58    89%
+simaas/tests/test_service_rti_aws.py          494     51    90%
+---------------------------------------------------------------
+TOTAL                                       12743   1907    85%
