+++ simaas/tests/helpers/__init__.py
@@ -1,0 +1,1 @@
+# Test helpers package
