+++ simaas/tests/integration/processors/__init__.py
@@ -1,0 +1,1 @@
+# Processor integration tests package
