+++ simaas/tests/integration/services/__init__.py
@@ -1,0 +1,1 @@
+# Service integration tests package
